+++ packages/expo-router/src/standard-navigation/types.ts
@@ -62,6 +62,7 @@ export type StandardUseNavigationBuilderOptions<
 export interface StandardNavigatorCreatePropsFactoryDeps<State extends NavigationState> {
   state: State;
   dispatch: (action: NavigationAction) => void;
+  dispatchSync: (action: NavigationAction) => void;
   navigation: NavigationHelpers<ParamListBase>;
 }